backend/requirements.txt
fastapi
uvicorn[standard]
firebase-admin
google-genai
python-dotenv
livekit-api
